# Reduce allocations & speed up enum writer