--- a/src/AvroConvert/AvroObjectServices/Schemas/EnumSchema.cs
+++ b/src/AvroConvert/AvroObjectServices/Schemas/EnumSchema.cs
@@ -13,7 +13,7 @@
 // See the Apache Version 2.0 License for specific language governing
 // permissions and limitations under the License.
 
-/** Modifications copyright(C) 2020 Adrian Struga³a **/
+/** Modifications copyright(C) 2020 Adrian Strugaï¿½a **/
 
 using System;
 using System.Collections.Generic;
@@ -84,13 +84,11 @@ namespace SolTechnology.Avro.AvroObjectServices.Schemas
             }
         }
         
-        internal ReadOnlyCollection<string> Symbols => new ReadOnlyCollection<string>(this.symbols);
+        internal ReadOnlyCollection<string> Symbols => new (this.symbols);
 
-        internal int GetValueBySymbol(string symbol)
-        {
-            return this.symbolToValue[symbol];
-        }
-
+        internal bool TryGetSymbolValue(string symbol, out int value) =>
+            this.symbolToValue.TryGetValue(symbol, out value);
+        
         internal string GetSymbolByValue(int value)
         {
             return this.valueToSymbol[value];